requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-dotenv
requests
PyPDF2
python-docx
fuzzywuzzy
python-Levenshtein
fastapi
uvicorn
pydantic
schedule
pytest
pytest-mock